+++ platformio/commands/list.py
@@ -7,15 +7,26 @@ from platformio.platforms.base import PlatformFactory
 
 
 @click.command("list", short_help="List installed platforms")
-def cli():
+@click.option("--json-output", is_flag=True)
+def cli(json_output):
 
     installed_platforms = PlatformFactory.get_platforms(
         installed=True).keys()
     installed_platforms.sort()
 
+    data = []
     for platform in installed_platforms:
         p = PlatformFactory.newPlatform(platform)
-        click.echo("{name:<20} with packages: {pkgs}".format(
-            name=click.style(p.get_name(), fg="cyan"),
-            pkgs=", ".join(p.get_installed_packages())
-        ))
+        data.append({
+            "name": platform,
+            "packages": p.get_installed_packages()
+        })
+
+    if json_output:
+        click.echo(data)
+    else:
+        for item in data:
+            click.echo("{name:<20} with packages: {pkgs}".format(
+                name=click.style(item['name'], fg="cyan"),
+                pkgs=", ".join(item['packages'])
+            ))